reworking build for react package
