+++ Mezzurite.React/package.json
@@ -11,14 +11,16 @@
   "keywords": [
     "mezzurite-react"
   ],
-  "main": "./dist/index.js",
-  "module": "./dist-esm/index.js",
+  "main": "./dist/bundles/mezzurite-react.umd.js",
+  "module": "./esm/mezzurite-react.js",
   "publishConfig": {
     "access": "restricted"
   },
   "scripts": {
     "clean": "shx rm -rf browser dist dist-esm",
-    "build": "npm run clean && rollup -c"
+    "build": "npm run clean && node build.js",
+    "pack:lib": "npm run build && npm pack ./dist",
+    "publish:lib": "npm run build && npm publish ./dist"
   },
   "babel": {
     "presets": [
@@ -39,13 +41,19 @@
     "@babel/preset-typescript": "^7.1.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.10.0",
+    "eslint-config-defaults": "^9.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-react": "^7.11.1",
     "path": "^0.12.7",
     "rollup": "^0.65.2",
     "rollup-plugin-babel": "^4.0.3",
     "rollup-plugin-json": "^3.1.0",
     "rollup-plugin-node-resolve": "^3.4.0",
     "rollup-plugin-uglify-es": "0.0.1",
-    "shx": "^0.3.2"
+    "shx": "^0.3.2",
+    "uglify-js": "^3.4.9"
   },
   "dependencies": {
     "@microsoft/mezzurite-core": "^1.0.1",
