+++ Mezzurite.React/rollup.config.js
@@ -1,13 +1,12 @@
 import resolve from 'rollup-plugin-node-resolve';
 import babel from 'rollup-plugin-babel';
-import uglify from 'rollup-plugin-uglify-es';
 import json from 'rollup-plugin-json';
 
 export default [
 {
-    input: './src/index.js',
+    input: './src/mezzurite-react.js',
     output: {
-    dir: 'dist-esm',
+    dir: 'dist/esm',
     format: 'esm',
     chunkFileNames: "[name]-common.js",
     globals: {
@@ -24,39 +23,15 @@ export default [
         resolve(['.js', '.json']),
         babel({
             exclude: 'node_modules/**'
-        }),
-        uglify()
+        })
     ]
 },
 {
-    input: './src/index.js',
+    input: './src/mezzurite-react.js',
     output: {
-    dir: 'dist',
-    format: 'cjs',
-    chunkFileNames: "[name]-common.js",
-    globals: {
-        react: 'React',
-        'react-dom': 'ReactDOM'
-    }
-    },
-    experimentalCodeSplitting: true,
-    external: [
-        'react', 'react-dom', 'react-router', 'intersection-observer', '@microsoft/mezzurite-core'
-    ],
-    plugins: [
-        json(),
-        resolve(['.js', '.json']),
-        babel({
-            exclude: 'node_modules/**'
-        }),
-        uglify()
-    ]
-},
-{
-    input: './src/index.js',
-    output: {
-    file: 'browser/mezzurite.react.umd.js',
+    file: 'dist/bundles/mezzurite-react.umd.js',
     name: 'Mezzurite-React',
+    sourcemap: true,
     format: 'umd',
     globals: {
         react: 'React',
@@ -74,7 +49,6 @@ export default [
         babel({
             exclude: 'node_modules/**'
         }),
-        uglify()
     ]
 }
 ]
