--- Mezzurite.React/src/index.js
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export * from './with-mezzurite.hoc';
-export * from './with-mezzurite-router.hoc';
-export * from './state-provider.service';
-export * from './performance-utils-react.service';
